[x86_64] Separate targets from main arch file
